WIP: Implement export beta
TODO: test
TODO: actually, this will break when we introduce taggable tags, so
maybe it should be postponed
TODO: I actually dislike that in the DataDump struct, interpretation of
DataDump and Data are up to the implementation. I think the dump format
should be unified.
TODO: IDs in the exported JSON need to be obfuscated. We'll need some
in-memory of that dump data for importing anyway, so maybe it's better
to do that now.